More frame method tests
