+++ pandas/tests/frame/test_analytics.py
@@ -1,7 +1,6 @@
 from datetime import timedelta
 from decimal import Decimal
 import operator
-import warnings
 
 import numpy as np
 import pytest
@@ -257,278 +256,11 @@ def assert_bool_op_api(
         getattr(bool_frame_with_na, opname)(axis=1, bool_only=False)
 
 
+class TestDataFrameCount:
+    pass
+
+
 class TestDataFrameAnalytics:
-
-    # ---------------------------------------------------------------------
-    # Correlation and covariance
-
-    @td.skip_if_no_scipy
-    def test_corr_pearson(self, float_frame):
-        float_frame["A"][:5] = np.nan
-        float_frame["B"][5:10] = np.nan
-
-        self._check_method(float_frame, "pearson")
-
-    @td.skip_if_no_scipy
-    def test_corr_kendall(self, float_frame):
-        float_frame["A"][:5] = np.nan
-        float_frame["B"][5:10] = np.nan
-
-        self._check_method(float_frame, "kendall")
-
-    @td.skip_if_no_scipy
-    def test_corr_spearman(self, float_frame):
-        float_frame["A"][:5] = np.nan
-        float_frame["B"][5:10] = np.nan
-
-        self._check_method(float_frame, "spearman")
-
-    def _check_method(self, frame, method="pearson"):
-        correls = frame.corr(method=method)
-        expected = frame["A"].corr(frame["C"], method=method)
-        tm.assert_almost_equal(correls["A"]["C"], expected)
-
-    @td.skip_if_no_scipy
-    def test_corr_non_numeric(self, float_frame, float_string_frame):
-        float_frame["A"][:5] = np.nan
-        float_frame["B"][5:10] = np.nan
-
-        # exclude non-numeric types
-        result = float_string_frame.corr()
-        expected = float_string_frame.loc[:, ["A", "B", "C", "D"]].corr()
-        tm.assert_frame_equal(result, expected)
-
-    @td.skip_if_no_scipy
-    @pytest.mark.parametrize("meth", ["pearson", "kendall", "spearman"])
-    def test_corr_nooverlap(self, meth):
-        # nothing in common
-        df = DataFrame(
-            {
-                "A": [1, 1.5, 1, np.nan, np.nan, np.nan],
-                "B": [np.nan, np.nan, np.nan, 1, 1.5, 1],
-                "C": [np.nan, np.nan, np.nan, np.nan, np.nan, np.nan],
-            }
-        )
-        rs = df.corr(meth)
-        assert isna(rs.loc["A", "B"])
-        assert isna(rs.loc["B", "A"])
-        assert rs.loc["A", "A"] == 1
-        assert rs.loc["B", "B"] == 1
-        assert isna(rs.loc["C", "C"])
-
-    @td.skip_if_no_scipy
-    @pytest.mark.parametrize("meth", ["pearson", "spearman"])
-    def test_corr_constant(self, meth):
-        # constant --> all NA
-
-        df = DataFrame(
-            {
-                "A": [1, 1, 1, np.nan, np.nan, np.nan],
-                "B": [np.nan, np.nan, np.nan, 1, 1, 1],
-            }
-        )
-        rs = df.corr(meth)
-        assert isna(rs.values).all()
-
-    def test_corr_int(self):
-        # dtypes other than float64 #1761
-        df3 = DataFrame({"a": [1, 2, 3, 4], "b": [1, 2, 3, 4]})
-
-        df3.cov()
-        df3.corr()
-
-    @td.skip_if_no_scipy
-    def test_corr_int_and_boolean(self):
-        # when dtypes of pandas series are different
-        # then ndarray will have dtype=object,
-        # so it need to be properly handled
-        df = DataFrame({"a": [True, False], "b": [1, 0]})
-
-        expected = DataFrame(np.ones((2, 2)), index=["a", "b"], columns=["a", "b"])
-        for meth in ["pearson", "kendall", "spearman"]:
-
-            with warnings.catch_warnings(record=True):
-                warnings.simplefilter("ignore", RuntimeWarning)
-                result = df.corr(meth)
-            tm.assert_frame_equal(result, expected)
-
-    def test_corr_cov_independent_index_column(self):
-        # GH 14617
-        df = pd.DataFrame(np.random.randn(4 * 10).reshape(10, 4), columns=list("abcd"))
-        for method in ["cov", "corr"]:
-            result = getattr(df, method)()
-            assert result.index is not result.columns
-            assert result.index.equals(result.columns)
-
-    def test_corr_invalid_method(self):
-        # GH 22298
-        df = pd.DataFrame(np.random.normal(size=(10, 2)))
-        msg = "method must be either 'pearson', 'spearman', 'kendall', or a callable, "
-        with pytest.raises(ValueError, match=msg):
-            df.corr(method="____")
-
-    def test_cov(self, float_frame, float_string_frame):
-        # min_periods no NAs (corner case)
-        expected = float_frame.cov()
-        result = float_frame.cov(min_periods=len(float_frame))
-
-        tm.assert_frame_equal(expected, result)
-
-        result = float_frame.cov(min_periods=len(float_frame) + 1)
-        assert isna(result.values).all()
-
-        # with NAs
-        frame = float_frame.copy()
-        frame["A"][:5] = np.nan
-        frame["B"][5:10] = np.nan
-        result = float_frame.cov(min_periods=len(float_frame) - 8)
-        expected = float_frame.cov()
-        expected.loc["A", "B"] = np.nan
-        expected.loc["B", "A"] = np.nan
-
-        # regular
-        float_frame["A"][:5] = np.nan
-        float_frame["B"][:10] = np.nan
-        cov = float_frame.cov()
-
-        tm.assert_almost_equal(cov["A"]["C"], float_frame["A"].cov(float_frame["C"]))
-
-        # exclude non-numeric types
-        result = float_string_frame.cov()
-        expected = float_string_frame.loc[:, ["A", "B", "C", "D"]].cov()
-        tm.assert_frame_equal(result, expected)
-
-        # Single column frame
-        df = DataFrame(np.linspace(0.0, 1.0, 10))
-        result = df.cov()
-        expected = DataFrame(
-            np.cov(df.values.T).reshape((1, 1)), index=df.columns, columns=df.columns
-        )
-        tm.assert_frame_equal(result, expected)
-        df.loc[0] = np.nan
-        result = df.cov()
-        expected = DataFrame(
-            np.cov(df.values[1:].T).reshape((1, 1)),
-            index=df.columns,
-            columns=df.columns,
-        )
-        tm.assert_frame_equal(result, expected)
-
-    def test_corrwith(self, datetime_frame):
-        a = datetime_frame
-        noise = Series(np.random.randn(len(a)), index=a.index)
-
-        b = datetime_frame.add(noise, axis=0)
-
-        # make sure order does not matter
-        b = b.reindex(columns=b.columns[::-1], index=b.index[::-1][10:])
-        del b["B"]
-
-        colcorr = a.corrwith(b, axis=0)
-        tm.assert_almost_equal(colcorr["A"], a["A"].corr(b["A"]))
-
-        rowcorr = a.corrwith(b, axis=1)
-        tm.assert_series_equal(rowcorr, a.T.corrwith(b.T, axis=0))
-
-        dropped = a.corrwith(b, axis=0, drop=True)
-        tm.assert_almost_equal(dropped["A"], a["A"].corr(b["A"]))
-        assert "B" not in dropped
-
-        dropped = a.corrwith(b, axis=1, drop=True)
-        assert a.index[-1] not in dropped.index
-
-        # non time-series data
-        index = ["a", "b", "c", "d", "e"]
-        columns = ["one", "two", "three", "four"]
-        df1 = DataFrame(np.random.randn(5, 4), index=index, columns=columns)
-        df2 = DataFrame(np.random.randn(4, 4), index=index[:4], columns=columns)
-        correls = df1.corrwith(df2, axis=1)
-        for row in index[:4]:
-            tm.assert_almost_equal(correls[row], df1.loc[row].corr(df2.loc[row]))
-
-    def test_corrwith_with_objects(self):
-        df1 = tm.makeTimeDataFrame()
-        df2 = tm.makeTimeDataFrame()
-        cols = ["A", "B", "C", "D"]
-
-        df1["obj"] = "foo"
-        df2["obj"] = "bar"
-
-        result = df1.corrwith(df2)
-        expected = df1.loc[:, cols].corrwith(df2.loc[:, cols])
-        tm.assert_series_equal(result, expected)
-
-        result = df1.corrwith(df2, axis=1)
-        expected = df1.loc[:, cols].corrwith(df2.loc[:, cols], axis=1)
-        tm.assert_series_equal(result, expected)
-
-    def test_corrwith_series(self, datetime_frame):
-        result = datetime_frame.corrwith(datetime_frame["A"])
-        expected = datetime_frame.apply(datetime_frame["A"].corr)
-
-        tm.assert_series_equal(result, expected)
-
-    def test_corrwith_matches_corrcoef(self):
-        df1 = DataFrame(np.arange(10000), columns=["a"])
-        df2 = DataFrame(np.arange(10000) ** 2, columns=["a"])
-        c1 = df1.corrwith(df2)["a"]
-        c2 = np.corrcoef(df1["a"], df2["a"])[0][1]
-
-        tm.assert_almost_equal(c1, c2)
-        assert c1 < 1
-
-    def test_corrwith_mixed_dtypes(self):
-        # GH 18570
-        df = pd.DataFrame(
-            {"a": [1, 4, 3, 2], "b": [4, 6, 7, 3], "c": ["a", "b", "c", "d"]}
-        )
-        s = pd.Series([0, 6, 7, 3])
-        result = df.corrwith(s)
-        corrs = [df["a"].corr(s), df["b"].corr(s)]
-        expected = pd.Series(data=corrs, index=["a", "b"])
-        tm.assert_series_equal(result, expected)
-
-    def test_corrwith_index_intersection(self):
-        df1 = pd.DataFrame(np.random.random(size=(10, 2)), columns=["a", "b"])
-        df2 = pd.DataFrame(np.random.random(size=(10, 3)), columns=["a", "b", "c"])
-
-        result = df1.corrwith(df2, drop=True).index.sort_values()
-        expected = df1.columns.intersection(df2.columns).sort_values()
-        tm.assert_index_equal(result, expected)
-
-    def test_corrwith_index_union(self):
-        df1 = pd.DataFrame(np.random.random(size=(10, 2)), columns=["a", "b"])
-        df2 = pd.DataFrame(np.random.random(size=(10, 3)), columns=["a", "b", "c"])
-
-        result = df1.corrwith(df2, drop=False).index.sort_values()
-        expected = df1.columns.union(df2.columns).sort_values()
-        tm.assert_index_equal(result, expected)
-
-    def test_corrwith_dup_cols(self):
-        # GH 21925
-        df1 = pd.DataFrame(np.vstack([np.arange(10)] * 3).T)
-        df2 = df1.copy()
-        df2 = pd.concat((df2, df2[0]), axis=1)
-
-        result = df1.corrwith(df2)
-        expected = pd.Series(np.ones(4), index=[0, 0, 1, 2])
-        tm.assert_series_equal(result, expected)
-
-    @td.skip_if_no_scipy
-    def test_corrwith_spearman(self):
-        # GH 21925
-        df = pd.DataFrame(np.random.random(size=(100, 3)))
-        result = df.corrwith(df ** 2, method="spearman")
-        expected = Series(np.ones(len(result)))
-        tm.assert_series_equal(result, expected)
-
-    @td.skip_if_no_scipy
-    def test_corrwith_kendall(self):
-        # GH 21925
-        df = pd.DataFrame(np.random.random(size=(100, 3)))
-        result = df.corrwith(df ** 2, method="kendall")
-        expected = Series(np.ones(len(result)))
-        tm.assert_series_equal(result, expected)
 
     # ---------------------------------------------------------------------
     # Reductions
@@ -1451,218 +1183,6 @@ class TestDataFrameAnalytics:
         xpr = "Must specify 'axis' when aggregating by level."
         with pytest.raises(ValueError, match=xpr):
             getattr(df, method)(axis=None, level="out")
-
-    # ---------------------------------------------------------------------
-    # Rounding
-
-    def test_round(self):
-        # GH 2665
-
-        # Test that rounding an empty DataFrame does nothing
-        df = DataFrame()
-        tm.assert_frame_equal(df, df.round())
-
-        # Here's the test frame we'll be working with
-        df = DataFrame({"col1": [1.123, 2.123, 3.123], "col2": [1.234, 2.234, 3.234]})
-
-        # Default round to integer (i.e. decimals=0)
-        expected_rounded = DataFrame({"col1": [1.0, 2.0, 3.0], "col2": [1.0, 2.0, 3.0]})
-        tm.assert_frame_equal(df.round(), expected_rounded)
-
-        # Round with an integer
-        decimals = 2
-        expected_rounded = DataFrame(
-            {"col1": [1.12, 2.12, 3.12], "col2": [1.23, 2.23, 3.23]}
-        )
-        tm.assert_frame_equal(df.round(decimals), expected_rounded)
-
-        # This should also work with np.round (since np.round dispatches to
-        # df.round)
-        tm.assert_frame_equal(np.round(df, decimals), expected_rounded)
-
-        # Round with a list
-        round_list = [1, 2]
-        with pytest.raises(TypeError):
-            df.round(round_list)
-
-        # Round with a dictionary
-        expected_rounded = DataFrame(
-            {"col1": [1.1, 2.1, 3.1], "col2": [1.23, 2.23, 3.23]}
-        )
-        round_dict = {"col1": 1, "col2": 2}
-        tm.assert_frame_equal(df.round(round_dict), expected_rounded)
-
-        # Incomplete dict
-        expected_partially_rounded = DataFrame(
-            {"col1": [1.123, 2.123, 3.123], "col2": [1.2, 2.2, 3.2]}
-        )
-        partial_round_dict = {"col2": 1}
-        tm.assert_frame_equal(df.round(partial_round_dict), expected_partially_rounded)
-
-        # Dict with unknown elements
-        wrong_round_dict = {"col3": 2, "col2": 1}
-        tm.assert_frame_equal(df.round(wrong_round_dict), expected_partially_rounded)
-
-        # float input to `decimals`
-        non_int_round_dict = {"col1": 1, "col2": 0.5}
-        with pytest.raises(TypeError):
-            df.round(non_int_round_dict)
-
-        # String input
-        non_int_round_dict = {"col1": 1, "col2": "foo"}
-        with pytest.raises(TypeError):
-            df.round(non_int_round_dict)
-
-        non_int_round_Series = Series(non_int_round_dict)
-        with pytest.raises(TypeError):
-            df.round(non_int_round_Series)
-
-        # List input
-        non_int_round_dict = {"col1": 1, "col2": [1, 2]}
-        with pytest.raises(TypeError):
-            df.round(non_int_round_dict)
-
-        non_int_round_Series = Series(non_int_round_dict)
-        with pytest.raises(TypeError):
-            df.round(non_int_round_Series)
-
-        # Non integer Series inputs
-        non_int_round_Series = Series(non_int_round_dict)
-        with pytest.raises(TypeError):
-            df.round(non_int_round_Series)
-
-        non_int_round_Series = Series(non_int_round_dict)
-        with pytest.raises(TypeError):
-            df.round(non_int_round_Series)
-
-        # Negative numbers
-        negative_round_dict = {"col1": -1, "col2": -2}
-        big_df = df * 100
-        expected_neg_rounded = DataFrame(
-            {"col1": [110.0, 210, 310], "col2": [100.0, 200, 300]}
-        )
-        tm.assert_frame_equal(big_df.round(negative_round_dict), expected_neg_rounded)
-
-        # nan in Series round
-        nan_round_Series = Series({"col1": np.nan, "col2": 1})
-
-        # TODO(wesm): unused?
-        expected_nan_round = DataFrame(  # noqa
-            {"col1": [1.123, 2.123, 3.123], "col2": [1.2, 2.2, 3.2]}
-        )
-
-        with pytest.raises(TypeError):
-            df.round(nan_round_Series)
-
-        # Make sure this doesn't break existing Series.round
-        tm.assert_series_equal(df["col1"].round(1), expected_rounded["col1"])
-
-        # named columns
-        # GH 11986
-        decimals = 2
-        expected_rounded = DataFrame(
-            {"col1": [1.12, 2.12, 3.12], "col2": [1.23, 2.23, 3.23]}
-        )
-        df.columns.name = "cols"
-        expected_rounded.columns.name = "cols"
-        tm.assert_frame_equal(df.round(decimals), expected_rounded)
-
-        # interaction of named columns & series
-        tm.assert_series_equal(df["col1"].round(decimals), expected_rounded["col1"])
-        tm.assert_series_equal(df.round(decimals)["col1"], expected_rounded["col1"])
-
-    def test_numpy_round(self):
-        # GH 12600
-        df = DataFrame([[1.53, 1.36], [0.06, 7.01]])
-        out = np.round(df, decimals=0)
-        expected = DataFrame([[2.0, 1.0], [0.0, 7.0]])
-        tm.assert_frame_equal(out, expected)
-
-        msg = "the 'out' parameter is not supported"
-        with pytest.raises(ValueError, match=msg):
-            np.round(df, decimals=0, out=df)
-
-    def test_numpy_round_nan(self):
-        # See gh-14197
-        df = Series([1.53, np.nan, 0.06]).to_frame()
-        with tm.assert_produces_warning(None):
-            result = df.round()
-        expected = Series([2.0, np.nan, 0.0]).to_frame()
-        tm.assert_frame_equal(result, expected)
-
-    def test_round_mixed_type(self):
-        # GH 11885
-        df = DataFrame(
-            {
-                "col1": [1.1, 2.2, 3.3, 4.4],
-                "col2": ["1", "a", "c", "f"],
-                "col3": date_range("20111111", periods=4),
-            }
-        )
-        round_0 = DataFrame(
-            {
-                "col1": [1.0, 2.0, 3.0, 4.0],
-                "col2": ["1", "a", "c", "f"],
-                "col3": date_range("20111111", periods=4),
-            }
-        )
-        tm.assert_frame_equal(df.round(), round_0)
-        tm.assert_frame_equal(df.round(1), df)
-        tm.assert_frame_equal(df.round({"col1": 1}), df)
-        tm.assert_frame_equal(df.round({"col1": 0}), round_0)
-        tm.assert_frame_equal(df.round({"col1": 0, "col2": 1}), round_0)
-        tm.assert_frame_equal(df.round({"col3": 1}), df)
-
-    def test_round_issue(self):
-        # GH 11611
-
-        df = pd.DataFrame(
-            np.random.random([3, 3]),
-            columns=["A", "B", "C"],
-            index=["first", "second", "third"],
-        )
-
-        dfs = pd.concat((df, df), axis=1)
-        rounded = dfs.round()
-        tm.assert_index_equal(rounded.index, dfs.index)
-
-        decimals = pd.Series([1, 0, 2], index=["A", "B", "A"])
-        msg = "Index of decimals must be unique"
-        with pytest.raises(ValueError, match=msg):
-            df.round(decimals)
-
-    def test_built_in_round(self):
-        # GH 11763
-        # Here's the test frame we'll be working with
-        df = DataFrame({"col1": [1.123, 2.123, 3.123], "col2": [1.234, 2.234, 3.234]})
-
-        # Default round to integer (i.e. decimals=0)
-        expected_rounded = DataFrame({"col1": [1.0, 2.0, 3.0], "col2": [1.0, 2.0, 3.0]})
-        tm.assert_frame_equal(round(df), expected_rounded)
-
-    def test_round_nonunique_categorical(self):
-        # See GH21809
-        idx = pd.CategoricalIndex(["low"] * 3 + ["hi"] * 3)
-        df = pd.DataFrame(np.random.rand(6, 3), columns=list("abc"))
-
-        expected = df.round(3)
-        expected.index = idx
-
-        df_categorical = df.copy().set_index(idx)
-        assert df_categorical.shape == (6, 3)
-        result = df_categorical.round(3)
-        assert result.shape == (6, 3)
-
-        tm.assert_frame_equal(result, expected)
-
-    def test_round_interval_category_columns(self):
-        # GH 30063
-        columns = pd.CategoricalIndex(pd.interval_range(0, 2))
-        df = DataFrame([[0.66, 1.1], [0.3, 0.25]], columns=columns)
-
-        result = df.round()
-        expected = DataFrame([[1.0, 1.0], [0.0, 0.0]], columns=columns)
-        tm.assert_frame_equal(result, expected)
 
     # ---------------------------------------------------------------------
     # Matrix-like
